Fixup sizes of network packet fields
With some additional hackery to enable UDP, this is enough for linux <-> windows multiplayer

+++ CODE/connect.h
@@ -121,7 +121,7 @@ PacketID:		This is a unique numerical ID for this packet.  The Connection
 typedef struct {
 	unsigned short MagicNumber;
 	unsigned char Code;
-	unsigned long PacketID;
+	uint32_t PacketID;
 } CommHeaderType;
 
 
@@ -262,13 +262,13 @@ class ConnectionClass
 		than this, we know is a resend.  Anything newer than this MUST be lying
 		around in the Queue for us to detect it as a resend.
 		.....................................................................*/
-		unsigned long LastSeqID;
+		uint32_t LastSeqID;
 
 		/*.....................................................................
 		This is the ID of the PACKET_DATA_ACK packet we read last; it ensures
 		that the application reads that type of packet in order.
 		.....................................................................*/
-		unsigned long LastReadID;
+		uint32_t LastReadID;
 
 		/*.....................................................................
 		Names of all packet commands
